YARN-10963. Split TestCapacityScheduler by test categories. Contributed by Tamas Domok
